ref: bundle exec-harness, memtrack, samply as `codspeed tool` subcommands
Move samply from a flat hidden `codspeed samply` re-exec target to
`codspeed tool samply` under a new hidden `tool` subcommand group, and
add `codspeed tool exec-harness` + `codspeed tool memtrack` alongside it.

Both helpers were previously shipped as standalone binaries downloaded
at runtime via pinned installer scripts. They now live entirely inside
the main `codspeed` binary: the orchestrator re-execs `current_exe tool
exec-harness …` / `tool memtrack track …` instead of relying on PATH,
the `BinaryPin` entries + `ensure_binary_installed` flow are removed,
and `src/binary_installer/` is deleted (only mongo-tracer remained, and
it uses its own download path).

Each crate keeps a `run_cli(argv)` entry shared between its standalone
`[[bin]]` (kept for development) and the bundled subcommand, so the two
paths can't drift. The standalone binaries are no longer released.

memtrack's eBPF code is Linux-only — the main crate enables the `ebpf`
feature only via `[target.'cfg(target_os = "linux")'.dependencies]`. On
macOS, `codspeed tool memtrack` parses but bails with a clear error
instead of a clap "unknown subcommand" message, keeping the CLI surface
uniform across platforms.

The `tool` group is hidden and disables its own help, so the root CLI's
global flags (`--config`, `--config-name`, `--setup-cache-dir`) no longer
leak into wrapped binaries' `--help` output; `--help` flows through to
samply/exec-harness/memtrack themselves. `Tool` short-circuits in
`cli::run()` before any config loading so a broken
`~/.config/codspeed/config.yaml` doesn't block re-exec.
