+++ src/executor/memory/mod.rs
@@ -1,2 +1,1 @@
 pub mod executor;
-pub(crate) mod setup;
