+++ src/executor/orchestrator.rs
@@ -1,7 +1,5 @@
 use super::{ExecutionContext, ExecutorName, get_executor_from_mode, run_executor};
 use crate::api_client::CodSpeedAPIClient;
-use crate::binary_installer::ensure_binary_installed;
-use crate::binary_pins::{self, PinnedBinary};
 use crate::cli::exec::multi_targets;
 use crate::cli::run::logger::Logger;
 use crate::executor::config::BenchmarkTarget;
@@ -17,9 +15,6 @@ use crate::upload::{UploadResult, upload};
 use serde_json::Value;
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
-
-pub const EXEC_HARNESS_COMMAND: &str = "exec-harness";
-pub const EXEC_HARNESS_VERSION: &str = binary_pins::EXEC_HARNESS_VERSION;
 
 /// Shared orchestration state created once per CLI invocation.
 ///
@@ -83,13 +78,6 @@ impl Orchestrator {
             .collect();
 
         if !exec_targets.is_empty() {
-            ensure_binary_installed(
-                EXEC_HARNESS_COMMAND,
-                EXEC_HARNESS_VERSION,
-                PinnedBinary::ExecHarnessInstaller,
-            )
-            .await?;
-
             let pipe_cmd = multi_targets::build_exec_targets_pipe_command(&exec_targets)?;
             let label = match exec_targets.as_slice() {
                 [BenchmarkTarget::Exec { command, .. }] => {
